+++ CHANGES.md
@@ -1,7 +1,8 @@
 unreleased
 ----------
 
-- Optimize and fix Linux package install (@MisterDA #153)
+- Optimize and fix Linux package install.
+  (@MisterDA #147, #151, #153, #154, review by @kit-ty-kate)
 - Switch to ocaml-opam/opam-repository-mingw#sunset for Windows images. (@MisterDA #152)
 - Use DockerHub user risvc64/ubuntu. (@MisterDA, #150)
 - Various LCU Updates (@mtelvers #144 #136 #135)